+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # tye-plugin Changelog
 
 ## [Unreleased]
+## [0.0.3]
 ### Added
 - New options in the run configuration to send logs, traces or specify tags.